Refactored database connection setup to use an interface than an object.
