+++ api/authentication/src/main/kotlin/uk/co/baconi/oauth/api/authentication/AuthenticationServer.kt
@@ -2,8 +2,7 @@ package uk.co.baconi.oauth.api.authentication
 
 import io.ktor.server.routing.*
 import uk.co.baconi.oauth.api.common.CommonModule.common
-import uk.co.baconi.oauth.api.common.DatabaseFactory.customerCredentialDatabase
-import uk.co.baconi.oauth.api.common.DatabaseFactory.customerStatusDatabase
+import uk.co.baconi.oauth.api.common.DatabaseModule
 import uk.co.baconi.oauth.api.common.TestUserModule
 import uk.co.baconi.oauth.api.common.authentication.CustomerAuthenticationService
 import uk.co.baconi.oauth.api.common.authentication.CustomerCredentialRepository
@@ -13,7 +12,7 @@ import uk.co.baconi.oauth.api.common.embeddedCommonServer
 /**
  * Start a server for just Authentication requests
  */
-internal object AuthenticationServer : AuthenticationRoute, TestUserModule {
+internal object AuthenticationServer : DatabaseModule, AuthenticationRoute, TestUserModule {
 
     override val customerCredentialRepository = CustomerCredentialRepository(customerCredentialDatabase)
     override val customerStatusRepository = CustomerStatusRepository(customerStatusDatabase)
