+++ api/authorisation/src/main/kotlin/uk/co/baconi/oauth/api/authorisation/AuthorisationServer.kt
@@ -2,8 +2,7 @@ package uk.co.baconi.oauth.api.authorisation
 
 import io.ktor.server.routing.*
 import uk.co.baconi.oauth.api.common.CommonModule.common
-import uk.co.baconi.oauth.api.common.DatabaseFactory
-import uk.co.baconi.oauth.api.common.DatabaseFactory.authorisationCodeDatabase
+import uk.co.baconi.oauth.api.common.DatabaseModule
 import uk.co.baconi.oauth.api.common.authorisation.AuthorisationCodeRepository
 import uk.co.baconi.oauth.api.common.client.ClientConfigurationRepository
 import uk.co.baconi.oauth.api.common.embeddedCommonServer
@@ -11,7 +10,7 @@ import uk.co.baconi.oauth.api.common.embeddedCommonServer
 /**
  * Start a server for just Authorisation requests
  */
-internal object AuthorisationServer : AuthorisationRoute {
+internal object AuthorisationServer : DatabaseModule, AuthorisationRoute {
 
     private val authorisationCodeRepository = AuthorisationCodeRepository(authorisationCodeDatabase)
     override val authorisationCodeService = AuthorisationCodeService(authorisationCodeRepository)
