+++ api/token-introspection/src/main/kotlin/uk/co/baconi/oauth/api/token/introspection/IntrospectionServer.kt
@@ -3,7 +3,7 @@ package uk.co.baconi.oauth.api.token.introspection
 import io.ktor.server.routing.*
 import uk.co.baconi.oauth.api.common.AuthenticationModule
 import uk.co.baconi.oauth.api.common.CommonModule.common
-import uk.co.baconi.oauth.api.common.DatabaseFactory.accessTokenDatabase
+import uk.co.baconi.oauth.api.common.DatabaseModule
 import uk.co.baconi.oauth.api.common.TestAccessTokenModule
 import uk.co.baconi.oauth.api.common.client.ClientConfigurationRepository
 import uk.co.baconi.oauth.api.common.client.ClientSecretRepository
@@ -15,7 +15,7 @@ import uk.co.baconi.oauth.api.common.token.AccessTokenService
 /**
  * Start a server for just Introspection requests
  */
-internal object IntrospectionServer : AuthenticationModule, IntrospectionRoute, TestAccessTokenModule {
+internal object IntrospectionServer : AuthenticationModule, DatabaseModule, IntrospectionRoute, TestAccessTokenModule {
 
     private val accessTokenRepository = AccessTokenRepository(accessTokenDatabase)
     override val accessTokenService = AccessTokenService(accessTokenRepository)
