+++ api/user-info/src/main/kotlin/uk/co/baconi/oauth/api/user/info/UserInfoServer.kt
@@ -3,7 +3,7 @@ package uk.co.baconi.oauth.api.user.info
 import io.ktor.server.routing.*
 import uk.co.baconi.oauth.api.common.AuthenticationModule
 import uk.co.baconi.oauth.api.common.CommonModule.common
-import uk.co.baconi.oauth.api.common.DatabaseFactory.accessTokenDatabase
+import uk.co.baconi.oauth.api.common.DatabaseModule
 import uk.co.baconi.oauth.api.common.TestAccessTokenModule
 import uk.co.baconi.oauth.api.common.client.ClientConfigurationRepository
 import uk.co.baconi.oauth.api.common.client.ClientSecretRepository
@@ -16,7 +16,7 @@ import uk.co.baconi.oauth.api.common.token.AccessTokenService
 /**
  * Start a server for just Introspection requests
  */
-internal object UserInfoServer : AuthenticationModule, UserInfoRoute, TestAccessTokenModule {
+internal object UserInfoServer : AuthenticationModule, DatabaseModule, UserInfoRoute, TestAccessTokenModule {
 
     private val accessTokenRepository = AccessTokenRepository(accessTokenDatabase)
     override val accessTokenService = AccessTokenService(accessTokenRepository)
